New mimetypes gone into effect today - this commit changes all old mimetypes over to the new mimetypes spec as descri...
Original commit message from CVS:
New mimetypes gone into effect today - this commit changes all old mimetypes over to the new mimetypes spec as described in the previous commit's document. Note: some plugins will break, some pipelines will break, expect HEAD to be broken or at least not 100% working for a few days, but don't forget to report bugs

+++ gst/videotestsrc/gstvideotestsrc.h
@@ -65,7 +65,7 @@ struct _GstVideotestsrc {
   gint64 timestamp;
   gint64 interval;
   gint bpp;
-  int rate;
+  gfloat rate;
   int type;
   GstClock *clock;
